+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.18.1 - Qol & Polish Patches
 
+### Features
+- Added a requeue button to the party dequeued message and /rfurequeue command.
+
 ### Fixes
 - Fixed date input in /rfudrophistory add not working.
 - Made hud not be hidden while on inventory